Update id fields to Long type

+++ src/main/java/com/example/reporting/repository/DepartmentRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.reporting.entity.Department;
 
 @Repository
-public interface DepartmentRepository extends JpaRepository<Department, Integer>{
+public interface DepartmentRepository extends JpaRepository<Department, Long>{
     
 }
